+++ src/libcollections/str.rs
@@ -1456,7 +1456,7 @@ impl str {
     /// characters be treated as 1 column (i.e., `is_cjk = false`) if the locale is unknown.
     #[unstable(feature = "unicode",
                reason = "this functionality may only be provided by libunicode")]
-    pub fn width(&self, is_cjk: bool) -> usize {
+    pub fn width(&self, is_cjk: bool) -> u32 {
         UnicodeStr::width(&self[..], is_cjk)
     }
